fixed URL parsing bug
