+++ bbot/core/helpers/__init__.py
@@ -1,3 +1,4 @@
 from .url import *
 from .misc import *
 from . import regexes
+from . import validators
