+++ bbot/core/helpers/url.py
@@ -1,10 +1,9 @@
-import re
 import uuid
 import logging
 from contextlib import suppress
 from urllib.parse import urlparse, parse_qs, urlencode, ParseResult
 
-from .punycode import smart_decode_punycode
+from .regexes import double_slash_regex
 
 
 log = logging.getLogger("bbot.core.helpers.url")
@@ -59,44 +58,6 @@ def param_type(p):
     return 3
 
 
-double_slash_regex = re.compile(r"/{2,}")
-
-
-def clean_url(url):
-    """
-    Remove query string and fragment, lowercase netloc, remove redundant port
-
-    http://evilcorp.com:80 --> http://evilcorp.com/
-    http://eViLcORp.com/ --> http://evilcorp.com/
-    http://evilcorp.com/api?user=bob#place --> http://evilcorp.com/api
-    """
-    parsed = parse_url(url)
-    parsed = parsed._replace(netloc=str(parsed.netloc).lower(), fragment="", query="")
-    try:
-        scheme = parsed.scheme
-    except ValueError:
-        scheme = "https"
-    try:
-        port = parsed.port
-    except ValueError:
-        port = 80 if scheme == "http" else 443
-    # remove ports if they're redundant
-    if (scheme == "http" and port == 80) or (scheme == "https" and port == 443):
-        hostname = parsed.hostname
-        # special case for IPv6 URLs
-        if parsed.netloc.startswith("["):
-            hostname = f"[{hostname}]"
-        # punycode
-        hostname = smart_decode_punycode(hostname)
-        parsed = parsed._replace(netloc=hostname)
-    # normalize double slashes
-    parsed = parsed._replace(path=double_slash_regex.sub("/", parsed.path))
-    # append / if path is empty
-    if parsed.path == "":
-        parsed = parsed._replace(path="/")
-    return parsed
-
-
 def hash_url(url):
     parsed = parse_url(url)
     parsed = parsed._replace(fragment="", query="")
@@ -112,37 +73,6 @@ def hash_url(url):
             hash_segment.append("")
         to_hash.append(tuple(hash_segment))
     return hash(tuple(to_hash))
-
-
-def collapse_urls(urls, threshold=10):
-    """
-    Smartly dedupe suspiciously-similar URLs like these:
-        - http://evilcorp.com/user/11111/info
-        - http://evilcorp.com/user/2222/info
-        - http://evilcorp.com/user/333/info
-        - http://evilcorp.com/user/44/info
-        - http://evilcorp.com/user/5/info
-
-    Useful for cleaning large lists of garbage-riddled URLs from sources like wayback
-    """
-    url_hashes = {}
-    for url in urls:
-        new_url = clean_url(url)
-        url_hash = hash_url(new_url)
-        try:
-            url_hashes[url_hash].add(new_url)
-        except KeyError:
-            url_hashes[url_hash] = {
-                new_url,
-            }
-
-    for url_hash, new_urls in url_hashes.items():
-        # if the number of URLs exceeds the threshold
-        if len(new_urls) > threshold:
-            # yield only one
-            yield next(iter(new_urls))
-        else:
-            yield from new_urls
 
 
 def url_depth(url):
